перенес reject/approve Passport в отдельный класс чтобы соблюдался srp
